Templates: fix the content type of templates Templating is weird af in Rocket. Look into 3a541ae for more. The content-type of the rendering is determined by the extension of the template name.
But renaming these templates would break building the project for
development because it finds a phony template there. So the trick is to
default the `template_dir` to `args.upload` because it should never
interfere and will be always present.
This also fixes the hax in Dockerfile by making it the default.

Signed-off-by: Gunwant Jain <mail@wantguns.dev>